Begining of bodyframe acceleration calculation for OpenRocket API.
At the moment, the code seems to be doing something reasonable, but it
has not been tested other than plotting the output on a single flight
file.  There needs to be quite a bit of testing done.  Also, it is
outputing invalid values after apogee, which is probably a result of the
OpenRocket file.  Need to see if there is a way to extend the available
data beyond that point.