Added to the 3D geometry module.
Planes can now be created from three points and lines from two points.
Additionally, planes and lines can be intersected (the function will
return None if the intersection is not defined, that is, the line and
plane are parallel).  Additionally, there are some small functions to
plot a point (numpy.array), line, and plane on a matplotlib axes in 3D.