Added new ECEF->LLH method (currently inaccurate) and did work on splashdown prediction.
The current splashdown prediction code is complete to the point of
generating what seem to be accurate predictions, however, those
predictions are returned through a side-channel (for debugging) and the
overall code is far from clean.  Much more work needs to be done to make
the code reliable and also to make the control parameters more
accessible.  Finally, the actual weighted mean/covariance calculation
needs to be completed correctly to return predictions.