Various changes, mostly additional splashdown work.
The GPS error model was made to have some additional noise control
options and the splashdown prediction code has added features to plot
the estimations through the course of the flight.  More work needs to be
done to clarify the algorighm and clean up the code.  Eventually, it
might be necessary for the prediction to be moved into C++ on the final
(Raspberry Pi-based base station), so any cleanup that can be done now
is a good thing.