Initial commit of the flightsim engine package.
This commit provides a directory for the package and contains the first
several modules that need to be used for the engine.  Most of these
files were simply copied from the original rocketry test directory and
should be preferred going forward.  For upcoming commits, in fact, all
of the useful files should be moved into the flightsim package, removed
from the original directory, and all of the existing tests rewritten to
use the package.